backend/requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
pydantic==2.10.3
python-dotenv==1.0.1
httpx==0.28.1
google-generativeai==0.8.3

